Fix indentation
It’s really important to spend the extra time necessary to make sure
that everything is indented properly.

Properly indented code is easier for everyone to read and understand,
including yourself.

Also, as you become more used to reading code that is indented
properly, you being to rely on it to tell you certain things about the
code without your having to read it. Sometimes, when code is indented
strangely, it can actually lie to the reader and make them think
something else is happening.