Initialize teams with 0 wins and 0 losses
Much of your code is pretty complex and I’m having a little difficulty
coming up with a clear path for refactoring. I’m going to walk through
an alternative approach to the problem starting from the beginning.

Here we’re going through the `score_data` array and creating a new hash
for each team that is a part of it, initialized with 0 wins and 0
losses.

This will result in a `@teams ` array that looks like this:

```ruby
[
  {:name=>"Patriots", :wins=>0, :losses=>0},
  {:name=>"Broncos", :wins=>0, :losses=>0},
  {:name=>"Broncos", :wins=>0, :losses=>0},
  {:name=>"Colts", :wins=>0, :losses=>0},
  {:name=>"Patriots", :wins=>0, :losses=>0},
  {:name=>"Colts", :wins=>0, :losses=>0},
  {:name=>"Steelers", :wins=>0, :losses=>0},
  {:name=>"Patriots", :wins=>0, :losses=>0}
]
```

You'll notice that we now have a hash in the `@teams` array for each of
the teams mentioned in the `score_data`. The one issue that we have is
that we have duplicate entries. We'll fix this in the next step.