Record game results
Here we use `Enumerable#find` to find each team by checking the value of
the team's name against either the home_team or the away_team based on
which team is won.

`#find` basically iterates over each value in the array, in our
case the `@teams`, until the code inside the block returns true. Once the
block returns true, the item that is currently being iterated over will
be returned.

After we find both the winning and losing team, update their values for
wins or losses depending on whether they won or lost.

At this point, our `@teams` array will look like this:

```ruby
[
  {:name=>"Patriots", :wins=>3, :losses=>0},
  {:name=>"Broncos", :wins=>1, :losses=>1},
  {:name=>"Colts", :wins=>0, :losses=>2},
  {:name=>"Steelers", :wins=>0, :losses=>1}
]
```

The only thing left to do in order to get this how we want it is to sort
the `@teams` array by `wins` and `losses`. We'll do that in the next
step.