Sort teams by wins and losses
Here we're using `Enumerable#sort_by` to sort our `@teams` array of
hashes.

This works similar to the `#sort` method that you were using except you
it's easier to sort by multiple values. Here we're saying that we want
to sort our teams by the wins, `team[:wins]`, and the losses,
`team[:losses]`.

We put a make the wins negative because we want higher values to be listed
first, and sort will sort by the given values in ASCENDING order, so making
the wins negative will make those wiht higher scores have a lower value
(making them be listed first).

This will leave us with our `@teams` array looking like this:

```ruby
[
  {:name=>"Patriots", :wins=>3, :losses=>0},
  {:name=>"Broncos", :wins=>1, :losses=>1},
  {:name=>"Steelers", :wins=>0, :losses=>1},
  {:name=>"Colts", :wins=>0, :losses=>2}
]
```

Now our `@teams` array has the correct information and is sorted in the
way that we want it. Now we can update the view for displaying the
information.