Fix the view
Now we can move all of that code into a `before` block so that it’s run
before the code inside the `get ‘/leaderboard’ do` is run.

This will make sure that the `@teams` instance variable is available in
the view.

Now that our view is working with the `@teams` array of hashes, it’s
much easier/explicit to grab the values and display them.